Removed Message Generation from CommandsHandler so it only handle Commands.
Now created MessageGenerator folder to keep every MessageGenerator partial class.

Each MessageGenerator_ file contains a MessageGenerator class' part for its purpose.

The main MessageGenerator file is meant for common stuff between all Generators.

This should permit a better maintenance of Message Generation.
